gdb: move remote_target::start_remote variable to narrower scope
The wait_status variable is only used when the target is in in all-stop
mode.  We can therefore move it in the !target_is_non_stop scope.  That
lets us remove the assert in the else, that checks that the wait status
is not set.  If the variable doesn't exist in that scope, it pretty much
guarantees that it is not set.

gdb/ChangeLog:

	* remote.c (remote_target::start_remote): Move wait_status to
	narrower scope.

Change-Id: I30979135e3f4f36d04178baa67575c4e58d3b648